// ai-assistant/projects/ai-assistant/src/lib/enums/feedback.enum.ts
export enum FeedbackOptions {
  LIKE = 'like',
  DISLIKE = 'dislike',
}

export enum ReasonEnum {
  MisunderstoodQuestion = 'Misunderstood Question',
  MisleadingInformation = 'Misleading Information',
  IncompleteAnswer = 'Incomplete Answer',
  OffTopicResponse = 'Off-Topic Response',
  NotWhatIWasLookingFor = 'Not What I Was Looking For',
  NotUseful = 'Not Useful',
  Other = 'Other',
}

export enum FeedBackActions {
  Upvote = 'upvote',
  Downvote = 'downvote',
  DownvoteWithReason = 'downvoteWithReason',
}